chore(frontend): chore place ico and disable tokenization
