+++ src/sections/projects/project-owner/detailForms/TokenizationForm.tsx
@@ -11,9 +11,6 @@ import { Button, Grid, InputAdornment, InputLabel, MenuItem, Select, Stack, Text
 import { useFormik } from 'formik';
 import * as yup from 'yup';
 
-import DateRangePicker from 'rsuite/DateRangePicker';
-import 'rsuite/dist/rsuite-no-reset.min.css';
-
 // project imports
 import AnimateButton from 'components/@extended/AnimateButton';
 import axios from 'utils/axios';
@@ -49,10 +46,6 @@ export type Tokenization = {
   tokenizingPercentage?: number;
   offeringPercentage?: number;
   minimumInvestment?: number;
-  fundSTDate?: Date;
-  fundEDDate?: Date;
-  tradingSTDate?: Date;
-  tradingEDDate?: Date;
 };
 
 interface TokenizationFormProps {
@@ -61,23 +54,11 @@ interface TokenizationFormProps {
   projectId?: string;
   projectType: boolean;
 }
-type TradingDuration = [Date, Date];
-
-const currentDate = new Date();
 
 export default function TokenizationForm({ tokenization, setTokenization, projectId, projectType }: TokenizationFormProps) {
   const [isSubmitting, setSubmitting] = useState<boolean>(false);
   const router = useRouter();
   const { data: session } = useSession();
-  const [tradingDuration, setTradingDuration] = useState<TradingDuration>([
-    new Date(tokenization.tradingSTDate ? tokenization.tradingSTDate : currentDate),
-    new Date(tokenization.tradingEDDate ? tokenization.tradingEDDate : currentDate)
-  ]);
-
-  const [fundRaisingDuration, setFundRaisingDuration] = useState<TradingDuration>([
-    new Date(tokenization.fundSTDate ? tokenization.fundSTDate : currentDate),
-    new Date(tokenization.fundEDDate ? tokenization.fundEDDate : currentDate)
-  ]);
 
   const formik = useFormik({
     initialValues: {
@@ -107,22 +88,10 @@ export default function TokenizationForm({ tokenization, setTokenization, projec
         minimumInvestment: values.minimumInvestment
       };
 
-      let tokenResult;
-      if (projectType) {
-        tokenResult = {
-          ...tokenInfo,
-          tradingSTDate: tradingDuration[0],
-          tradingEDDate: tradingDuration[1],
-          fundSTDate: fundRaisingDuration[0],
-          fundEDDate: fundRaisingDuration[1]
-        };
-      } else tokenResult = { ...tokenInfo };
-
-      console.log('tokenResult-->', tokenResult);
       axios.defaults.headers.common = { Authorization: `bearer ${session?.token.accessToken as string}` };
 
       axios
-        .post(`/api/v1/project/${projectId}/tokenization`, tokenResult)
+        .post(`/api/v1/project/${projectId}/tokenization`, tokenInfo)
         .then(async (res) => {
           enqueueSnackbar('Tokenized successfully.', {
             variant: 'success',
@@ -335,38 +304,6 @@ export default function TokenizationForm({ tokenization, setTokenization, projec
               </Select>
             </Stack>
           </Grid>
-          {projectType && (
-            <>
-              <Grid item xs={12}>
-                <Stack spacing={0.5}>
-                  <InputLabel>Trading Duration *</InputLabel>
-                  <DateRangePicker
-                    id="tradingDuration"
-                    name="tradingDuration"
-                    size="lg"
-                    value={tradingDuration}
-                    readOnly={router.query.projectId !== 'add-shipping' && router.query.projectId !== 'add-ico'}
-                    // @ts-ignore
-                    onChange={setTradingDuration}
-                  />
-                </Stack>
-              </Grid>
-              <Grid item xs={12}>
-                <Stack spacing={0.5}>
-                  <InputLabel>Fundraising Duration *</InputLabel>
-                  <DateRangePicker
-                    id="tradingDuration"
-                    name="tradingDuration"
-                    size="lg"
-                    value={fundRaisingDuration}
-                    readOnly={router.query.projectId !== 'add-shipping' && router.query.projectId !== 'add-ico'}
-                    // @ts-ignore
-                    onChange={setFundRaisingDuration}
-                  />
-                </Stack>
-              </Grid>
-            </>
-          )}
 
           {(router.query.projectId === 'add-shipping' || router.query.projectId === 'add-ico') && (
             <Grid item xs={12}>
